+++ sdk/formrecognizer/Azure.AI.FormRecognizer/tests/samples/Sample2_RecognizeReceiptsFromFile.cs
@@ -29,8 +29,8 @@ namespace Azure.AI.FormRecognizer.Samples
             #region Snippet:FormRecognizerSampleRecognizeReceiptFileStream
             using (FileStream stream = new FileStream(receiptPath, FileMode.Open))
             {
-                Response<RecognizedReceiptCollection> receipts = await client.StartRecognizeReceipts(stream).WaitForCompletionAsync();
-                foreach (var receipt in receipts.Value)
+                RecognizedReceiptCollection receipts = await client.StartRecognizeReceipts(stream).WaitForCompletionAsync();
+                foreach (var receipt in receipts)
                 {
                     USReceipt usReceipt = receipt.AsUSReceipt();