+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.0.0-preview.3 (Unreleased)
 
+### Breaking changes
+
+- `RecognizeContentOperation` now returns a `FormPageCollection`.
+- `RecognizeReceiptsOperation` now returns a `RecognizedReceiptCollection`.
+- `RecognizeCustomFormsOperation` now returns a `RecognizedFormCollection`.
 
 ## 1.0.0-preview.2 (05-06-2020)